feat: CRUD operation for Task
refs #8

+++ service/src/main/java/su/itpro/tasktracker/model/entity/Category.java
@@ -21,11 +21,11 @@ import lombok.ToString;
 @NoArgsConstructor
 @Builder
 @Entity
-public class Category implements BaseEntity<Integer> {
+public class Category implements BaseEntity<Short> {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Short id;
 
   @Column(nullable = false, unique = true)
   private String name;
